requirements.txt
flask
flask-sqlalchemy
flask-wtf
wtforms-sqlalchemy
flask-bcrypt
flask-mail
